first part of tidying up
