+++ core/optical_element.py
@@ -21,78 +21,12 @@ along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 """
 
-import surfShape
-from material import ConstantIndexGlass
 
-import aperture
-import pupil
-from coordinates import LocalCoordinates
-
-#import inspector
-import numpy as np
-from optimize import ClassWithOptimizableVariables
-from optimize import OptimizableVariable
-
-from ray import RayPathNew, RayBundleNew
-
-from globalconstants import canonical_ex, canonical_ey, canonical_ez
-
-import uuid
+from core.localcoordinatestreebase import LocalCoordinatesTreeBase
+from ray import RayPath
 
 
-class CoordinateTreeBase(ClassWithOptimizableVariables):
-    """
-    Optical element base class for optical system and optical element, whereas
-    an optical system consists of many optical elements.
-    Implements functionality for local coordinate system tree and connection
-    checks.
-    
-    :param rootcoordinatesystem (LocalCoordinates object)
-    :param label (string)
-    :param *kwargs (key word arguments)
-    """
-    def __init__(self, rootcoordinatesystem, label="", **kwargs):
-        self.label = label
-        self.rootcoordinatesystem = rootcoordinatesystem
-        
-    def checkForRootConnection(self, lc):
-        """
-        Checks whether given local coordinate system is child of rootcoordinatesystem.
-        
-        :param lc (LocalCoordinates object)
-        
-        :return bool
-        """
-        allconnectedchildren = self.rootcoordinatesystem.returnConnectedChildren()        
-        return (lc in allconnectedchildren)
-            
-    def addLocalCoordinateSystem(self, lc, refname):
-        """
-        Adds local coordinate system as child to given reference.
-        
-        :param lc (LocalCoordinates object)
-        :param refname (string)
-        
-        :return lc        
-        """
-        allnames = self.rootcoordinatesystem.returnConnectedNames()
-       
-        if lc.name in allnames:
-            lc.name = str(uuid.uuid4())
-            
-        if refname not in allnames:
-            refname = self.rootcoordinatesystem.name
-        
-        self.rootcoordinatesystem.addChildToReference(refname, lc)
-            
-        return lc
-    
-    
-    
-        
-
-
-class OpticalElement(CoordinateTreeBase):
+class OpticalElement(LocalCoordinatesTreeBase):
     """
     Represents an optical element (volume with surface boundary and inner
     surfaces representing material boundaries)
@@ -168,7 +102,7 @@ class OpticalElement(CoordinateTreeBase):
         
         current_material = background_medium    
     
-        rpath = RayPathNew(raybundle)    
+        rpath = RayPath(raybundle)    
     
         for (surfkey, refract_flag, ordinary_flag) in sequence:
             
@@ -186,9 +120,9 @@ class OpticalElement(CoordinateTreeBase):
 
             if refract_flag:
                 current_material = self.findoutWhichMaterial(mnmat, pnmat, current_material)
-                rpath.appendRayBundle(current_material.refractNew(current_bundle, current_surface))
+                rpath.appendRayBundle(current_material.refract(current_bundle, current_surface))
             else:
-                rpath.appendRayBundle(current_material.reflectNew(current_bundle, current_surface))
+                rpath.appendRayBundle(current_material.reflect(current_bundle, current_surface))
                 
 
 
@@ -196,332 +130,3 @@ class OpticalElement(CoordinateTreeBase):
 
 
 
-class SurfaceNew(CoordinateTreeBase):
-    """
-    Represents a surface of an optical system.
-
-    :param shape: Shape of the surface. Calculates the intersection with rays. ( Shape object or child )
-    :param material: Material of the volume behind the surface. Calculates the refraction. ( Material object or child )
-    :param thickness: distance to next surface on the optical axis
-    """
-    def __init__(self, rootlc, shape=None, apert=None, label=""):
-        super(SurfaceNew, self).__init__(rootlc, label)
-
-        if shape is None:        
-            shape = surfShape.Conic(rootlc)
-        if apert is None:
-            apert = aperture.BaseAperture(rootlc)
-            
-        self.setShape(shape)
-        self.setAperture(apert)
-
-
-    def setAperture(self, apert):
-        """
-        Sets the shape object self.shap
-
-        :param shape: the new Shape object
-
-        :return self.shape: new Shape object
-        """
-        if self.checkForRootConnection(apert.lc):
-            self.__aperture = apert
-        else:
-            raise Exception("Aperture coordinate system should be connected to surface coordinate system")            
-       
-    def getAperture(self):
-        return self.__aperture
-        
-    aperture = property(getAperture, setAperture)
-
-
-    def setShape(self, shape):
-        """
-        Sets the shape object self.shap
-
-        :param shape: the new Shape object
-
-        :return self.shape: new Shape object
-        """
-        if self.checkForRootConnection(shape.lc):
-            self.__shape = shape
-        else:
-            raise Exception("Shape coordinate system should be connected to surface coordinate system")            
-        
-    def getShape(self):
-        return self.__shape
-        
-    shape = property(getShape, setShape)
-    
-
-    def intersect(self, raybundle, remove_rays_outside_aperture=True):
-        """
-        Calculates intersection from raybundle. Knows shape and aperture and
-        can remove rays due to aperture.
-        
-        :param raybundle (RayBundle object), gets changed!
-        """
-        
-        self.shape.intersectNew(raybundle)
-        
-        if remove_rays_outside_aperture:
-            globalintersection = raybundle.x[-1]
-            local_ap_intersection = self.aperture.lc.returnGlobalToLocalPoints(globalintersection)
-        
-            valid = self.aperture.arePointsInAperture(local_ap_intersection[0], local_ap_intersection[1])
-        
-            raybundle.valid[-1] = raybundle.valid[-1]*valid
-
-
-    def draw2d(self, ax, vertices=100, inyzplane = True, color="grey", plane_normal = canonical_ex, up = canonical_ey):
-        """
-        :param ax (Axis object)
-        :param vertices (int), vertices in xy for aperture sampling
-        :param inyzplane (bool), cuts globalpts in yz plane before projection on plane_normal
-        :param color (string), "red", "blue", "grey", "green", ...
-        :param plane_normal (1D numpy array of float), new x projection axis
-        :param up (1D numpy array), invariant y axis, z = x x y
-        
-        """
-
-
-        sizelimit = 1000.0
-        failsafevalue = 10.0        
-        if self.aperture == None:
-            effsemidia = failsafevalue
-        else:
-            if self.aperture.getTypicalDimension() <= sizelimit:
-                # TODO: maybe introduce aperture types Object and Image to distuingish from very large normal apertures
-                effsemidia = self.aperture.getTypicalDimension() #self.sdia.val if self.sdia.val < 10.0 else 10.0
-            else:
-                effsemidia = failsafevalue
-        
-        xl = effsemidia * np.linspace(-1, 1, num=vertices)
-        yl = effsemidia * np.linspace(-1, 1, num=vertices)
-        
-        X, Y = np.meshgrid(xl, yl)
-        x = X.flatten()
-        y = Y.flatten()
-        
-        isinap = np.array(self.aperture.arePointsInAperture(x, y))
-        xinap = x[isinap]        
-        yinap = y[isinap]
-        zinap = np.zeros_like(xinap)
-        
-        localpts_aperture = np.row_stack((xinap, yinap, zinap))
-        localpts_shape = self.shape.lc.returnOtherToActualPoints(localpts_aperture, self.aperture.lc)
-        
-        xinap_shape = localpts_shape[0, :]
-        yinap_shape = localpts_shape[1, :]        
-        zinap_shape = self.shape.getSag(xinap_shape, yinap_shape)
-        
-        localpts_shape = np.row_stack((xinap_shape, yinap_shape, zinap_shape))
-        localpts_surf = self.rootcoordinatesystem.returnOtherToActualPoints(localpts_shape, self.shape.lc)        
-        
-        # ebenenprojektion hier!        
-        
-        globalpts = self.rootcoordinatesystem.returnLocalToGlobalPoints(localpts_surf)
-
-        # doubled code begin (also in RayBundleNew.draw2d)
-        plane_normal = plane_normal/np.linalg.norm(plane_normal)
-        up = up/np.linalg.norm(up)
-
-        ez = np.cross(plane_normal, up)
-
-        (num_dims, num_rays) = np.shape(globalpts)
-
-        # arrange num_ray copies of simple vectors in appropriate form
-        plane_normal = np.column_stack((plane_normal for i in np.arange(num_rays)))
-        ez = np.column_stack((ez for i in np.arange(num_rays)))
-        up = np.column_stack((up for i in np.arange(num_rays)))
-        # doubled code (also in RayBundleNew.draw2d)
-
-        # doubled code (see ray.py)
-
-        # show surfaces with points in yz plane before pn-plane projection
-        # if false show full surface
-        if inyzplane:
-            inYZplane = np.abs(globalpts[0]) < 2*effsemidia/vertices
-            globalpts = globalpts[:, inYZplane]
-            plane_normal = plane_normal[:, inYZplane]
-            up = up[:, inYZplane]
-            ez = ez[:, inYZplane]
-
-        globalptsinplane = globalpts - np.sum(globalpts*plane_normal, axis=0)*plane_normal
-
-       
-        # calculate y-components
-        ypt = np.sum(globalptsinplane * up, axis=0)
-        # calculate z-components
-        zpt = np.sum(globalptsinplane * ez, axis=0)
-
-        # doubled code (see ray.py)
-
-        
-        #ax.plot(zinap+offset[1], yinap+offset[0], color)
-        ax.plot(zpt, ypt, color)
-        
-        
-        #self.shape.draw2d(ax, offset, vertices, color, self.aperture)
-
-    def getCentralCurvature(self, ray):
-        curvature = self.shape.getCentralCurvature()
-        # TODO: curvature at ray position
-        
-        return curvature
-        
-
-class OpticalSystemNew(CoordinateTreeBase):
-    """
-    Represents an optical system, consisting of several surfaces and materials inbetween.
-    """
-    def __init__(self, rootlc = None, matbackground = None, name = ""):
-        # TODO: rename variable name to label
-        """
-        Creates an optical system object. Initially, it contains 2 plane surfaces (object and image).
-
-
-        :param objectLC: local coordinate system of object (class LocalCoordinates).
-
-
-        """
-        if rootlc is None:        
-            rootlc = LocalCoordinates(name="global")
-        self.rootcoordinatesystem = rootlc
-
-        super(OpticalSystemNew, self).__init__(self.rootcoordinatesystem, label = name)
-        
-        if matbackground is None:
-            matbackground = ConstantIndexGlass(self.rootcoordinatesystem, 1.0)
-
-        self.material_background = matbackground # Background material        
-        self.elements = {}
-
-    def seqtrace(self, initialbundle, elementsequence): # [("elem1", [1, 3, 4]), ("elem2", [1,4,4]), ("elem1", [4, 3, 1])]
-        rpath = RayPathNew(initialbundle)
-        for (elem, subseq) in elementsequence:
-            rpath.appendRayPath(self.elements[elem].seqtrace(rpath.raybundles[-1], subseq, self.material_background)) 
-        return rpath
-            
-
-    def addElement(self, key, element):
-        """
-        Adds a new element (containing several surfaces) into the optical system.
-
-        :param key (string)        
-        :param element (optical element class)
-        """
-        if self.checkForRootConnection(element.rootcoordinatesystem):
-            self.elements[key] = element
-        else:
-            raise Exception("OpticalElement root should be connected to root of OpticalSystem")
-
-    def removeElement(self, key):
-        """
-        Removes an optical element from the optical system.
-
-        :param key (string)
-        """
-        # TODO: update of local coordinate references missing
-        if key in self.elements:        
-            self.elements.pop(key)
-
-
-    def getABCDMatrix(self, ray, firstSurfacePosition=0, lastSurfacePosition=-1):
-        """
-        Returns an ABCD matrix of the optical system.
-        The matrix is set up in geometric convention for (y, dy/dz) vectors.
-
-        The matrix contains:
-        - paraxial refraction from vacuum through the first surface
-        - paraxial propagation through the system
-        - paraxial refraction after the last surface into vacuum
-
-        :param firstSurfacePosition: Position of the first surface to consider (int).
-          Preset is 0 (object position).
-        :param lastSurfacePosition: Position of the last surface to consider (int).
-          Preset is -1 (image position)
-        :param ray: Ray bundle object.
-        :return abcd: ABCD matrix (2d numpy 2x2 matrix of float)
-        """
-
-        if lastSurfacePosition < 0:
-            lastSurfacePosition = self.getNumberOfSurfaces() - lastSurfacePosition - 3
-
-        abcd = [[1, 0], [0, 1]]
-
-        for i in np.arange(lastSurfacePosition - firstSurfacePosition + 1) + firstSurfacePosition:
-            abcd = np.dot(self.surfaces[i].getABCDMatrix(self.surfaces[i+1], ray), abcd)
-
-        return abcd
-
-
-
-    def getParaxialPupil(self, stopPosition, ray):
-        """
-        Returns the paraxially calculated pupil positions.
-
-        :param stopPosition: surface number of the surface that is defined as stop (int)
-        :param ray: Raybundle object
-
-        :return zen: entrance pupil position from object (float)
-        :return magen: entrance pupil magnificaction; entrance pupil diameter per stop diameter (float)
-        :return zex: exit pupil position from image (float)
-        :return magex: exit pupil magnificaction; exit pupil diameter per stop diameter (float)
-        """
-        abcdObjStop = self.getABCDMatrix(ray, 0, stopPosition - 1)  # object to stop
-
-        zen = abcdObjStop[0, 1] / abcdObjStop[0, 0]  # entrance pupil position from object
-        magen = 1.0 / abcdObjStop[0, 0]
-
-        abcdStopIm = self.getABCDMatrix(ray, stopPosition, -1)  # stop to image
-
-        zex = - abcdStopIm[0, 1] / abcdStopIm[1, 1]  # exit pupil position from image
-        magex = abcdStopIm[0, 0] - abcdStopIm[0, 1] * abcdStopIm[1, 0] / abcdStopIm[1, 1]
-
-        return zen, magen, zex, magex, abcdObjStop, abcdStopIm
-
-    def getEffectiveFocalLength(self, ray):
-        """
-        Returns the effective (paraxial) focal length of the system.
-
-        :param ray: Raybundle object
-        :return f: focal length (float)
-        """
-        abcd = self.getABCDMatrix(ray)
-        return -1.0 / abcd[1, 0]
-
-    def getParaxialMagnification(self, ray):
-        """
-        Returns the paraxial real space magnification of the system.
-        Before calculation, the image is shifted into paraxial   finite conjugate plane.
-
-        :param ray: Raybundle object
-        :return pmag: real space paraxial magnification (float)
-        """
-        abcd = self.getABCDMatrix(ray)
-        print abcd
-        return abcd[0, 0] - abcd[0, 1] * abcd[1, 0] / abcd[1, 1]
-
-
-    def draw2d(self, ax, vertices=100, color="grey"):
-        for (num, s) in enumerate(self.surfaces):
-            s.draw2d(ax, vertices=vertices, color=color)
-            
-
-
-if __name__ == "__main__":
-
-    # AC254-100-Ad 	25.4 	100.1 	97.1 	info 	62.8 	-45.7 	-128.2 	4.0 	2.5 	4.7 	N-BK7/SF5    
-    
-    os = OpticalSystemNew()
-    
-    lc1 = os.addLocalCoordinateSystem(LocalCoordinates(decz=10.0), refname=os.rootcoordinatesystem.name)
-    lc2 = os.addLocalCoordinateSystem(LocalCoordinates(decz=20.0), refname=lc1.name)
-    lc3 = os.addLocalCoordinateSystem(LocalCoordinates(decz=30.0), refname=lc2.name)
-    lc4 = os.addLocalCoordinateSystem(LocalCoordinates(decz=40.0), refname=lc3.name)
-    
-    lc5 = os.addLocalCoordinateSystem(LocalCoordinates(name="COM", decx=10.0, decy=5.0, decz=10.), refname=lc1.name)
-    
-    print(os.rootcoordinatesystem.pprint())
-
