+++ demo_doublet.py
@@ -21,6 +21,12 @@ along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 """
 
+"""
+Smoke demo for picture checking. This means the result of this file
+is compared with a fixed picture and may not change. That's part of the test.
+If anything breaks, the picture changes usually.
+"""
+
 
 import numpy as np
 import matplotlib.pyplot as plt
@@ -28,20 +34,22 @@ import matplotlib.pyplot as plt
 from core import raster
 from core import material
 from core import surfShape
-from core.optical_element import OpticalSystemNew, SurfaceNew, OpticalElement
-from core.ray import RayBundleNew
+from core.optical_element import OpticalElement
+from core.surface import Surface
+from core.optical_system import OpticalSystem
+from core.ray import RayBundle
 
 from core.aperture import CircularAperture
-from core.coordinates import LocalCoordinates
+from core.localcoordinates import LocalCoordinates
 
-from core.globalconstants import canonical_ex, canonical_ey
+from core.globalconstants import canonical_ey
 
 import math
 
 wavelength = 0.5876e-3
 
 # definition of optical system
-s = OpticalSystemNew() 
+s = OpticalSystem() 
 
 lc0 = s.addLocalCoordinateSystem(LocalCoordinates(name="stop", decz=0.0), refname=s.rootcoordinatesystem.name)
 lc1 = s.addLocalCoordinateSystem(LocalCoordinates(name="surf1", decz=-1.048), refname=lc0.name) # objectDist
@@ -50,11 +58,11 @@ lc3 = s.addLocalCoordinateSystem(LocalCoordinates(name="surf3", decz=2.5), refna
 lc4 = s.addLocalCoordinateSystem(LocalCoordinates(name="image", decz=97.2), refname=lc3.name)
 
 
-stopsurf = SurfaceNew(lc0)
-frontsurf = SurfaceNew(lc1, shape=surfShape.Conic(lc1, curv=1./62.8), apert=CircularAperture(lc1, 12.7))
-cementsurf = SurfaceNew(lc2, shape=surfShape.Conic(lc2, curv=-1./45.7), apert=CircularAperture(lc2, 12.7))
-rearsurf = SurfaceNew(lc3, shape=surfShape.Conic(lc3, curv=-1./128.2), apert=CircularAperture(lc3, 12.7))
-image = SurfaceNew(lc4)
+stopsurf = Surface(lc0)
+frontsurf = Surface(lc1, shape=surfShape.Conic(lc1, curv=1./62.8), apert=CircularAperture(lc1, 12.7))
+cementsurf = Surface(lc2, shape=surfShape.Conic(lc2, curv=-1./45.7), apert=CircularAperture(lc2, 12.7))
+rearsurf = Surface(lc3, shape=surfShape.Conic(lc3, curv=-1./128.2), apert=CircularAperture(lc3, 12.7))
+image = Surface(lc4)
 
 
 elem = OpticalElement(lc0, label="thorlabs_AC_254-100-A")
@@ -73,12 +81,10 @@ elem.addSurface("image", image, (None, None))
 
 s.addElement("AC254-100", elem)
 
-print(s.rootcoordinatesystem.pprint())
-
 rstobj = raster.MeridionalFan()
 (px, py) = rstobj.getGrid(20)
 
-rpup = 20. #11.43
+rpup = 11.43
 o = np.vstack((rpup*px, rpup*py, -5.*np.ones_like(px)))
 
 k = np.zeros_like(o)
@@ -93,24 +99,8 @@ sysseq = [("AC254-100", [("stop", True, True), ("front", True, True), ("cement",
 
 phi = 5.*math.pi/180.0
 
-initialbundle = RayBundleNew(x0=o, k0=k, Efield0=E0, wave=wavelength)
+initialbundle = RayBundle(x0=o, k0=k, Efield0=E0, wave=wavelength)
 r2 = s.seqtrace(initialbundle, sysseq)
-
-pilotbundle = RayBundleNew(
-                x0 = np.array([[0], [0], [0]]), 
-                k0 = np.array([[0], [2.*math.pi/wavelength*math.sin(phi)], [2.*math.pi/wavelength*math.cos(phi)]]), 
-                Efield0 = np.array([[1], [0], [0]]), wave=wavelength
-                )
-
-pilotray = s.seqtrace(pilotbundle, sysseq)
-
-#for (ind, r) in enumerate(pilotray.raybundles):
-#    print("pilot bundle %d" % (ind,))
-#    print(r.x)
-    
-#print("last coordinates")
-#print(r2.raybundles[-1].x[-1, :, :])
-
 
 fig = plt.figure(1)
 ax = fig.add_subplot(111)
@@ -122,16 +112,12 @@ phi = 0.#math.pi/4
 pn = np.array([math.cos(phi), 0, math.sin(phi)]) # canonical_ex
 up = canonical_ey
 
-print("drawing!")
 r2.draw2d(ax, color="blue", plane_normal=pn, up=up) 
-pilotray.draw2d(ax, color="darkgreen", plane_normal=pn, up=up)
 for e in s.elements.itervalues():
     for surfs in e.surfaces.itervalues():
         surfs.draw2d(ax, color="grey", vertices=50, plane_normal=pn, up=up) # try for phi=0.
         #surfs.draw2d(ax, color="grey", inyzplane=False, vertices=50, plane_normal=pn, up=up) # try for phi=pi/4
 
-#plots.drawLayout2d(ax, s, [pilotpath])
-#plots.drawLayout2d(ax, s, [r2])
 
 plt.show()
